faiss improvement:1 use recommended parameter
tmp
